Implements automatic wiki information lookup based on dbnames and siteinfo.
