+++ config/00-main.yaml
@@ -13,22 +13,3 @@ database:
 
 oauth:
   mw_uri: https://www.mediawiki.org/w/index.php
-
-wikis:
-  testwiki:
-    db_name: 'testwiki'
-    wiki_path: 'test.wikipedia.org'
-    lang_code: 'en'
-  fawiki:
-    db_name: 'fawiki'
-    wiki_path: 'fa.wikipedia.org'
-    lang_code: 'fa'
-    dir: rtl
-  enwiki:
-    db_name: 'enwiki'
-    wiki_path: 'en.wikipedia.org'
-    lang_code: 'en'
-  mediawikiwiki:
-    db_name: 'mediawikiwiki'
-    wiki_path: 'mediawiki.org'
-    lang_code: 'en'
