+++ wikilabels/wsgi/routes/ui.py
@@ -1,6 +1,7 @@
 from flask import render_template
 
 from .. import assets, preprocessors, responses
+from ...util import wikimedia
 from ..util import app_path, build_script_tags, build_style_tags, url_for
 
 
@@ -14,8 +15,6 @@ def configure(bp, config):
     @bp.route("/ui/<wiki>/")
     @preprocessors.debuggable
     def ui_wiki(wiki):
-        if wiki not in config['wikis']:
-            return responses.not_found('Wiki config not found')
         script_tags = '<script src="{0}"></script>' \
             .format(app_path('/gadget/' + wiki + '/mediawiki.js',
                     config))
@@ -26,6 +25,6 @@ def configure(bp, config):
                                script_tags=script_tags,
                                style_tags=style_tags,
                                server_root=url_for("", config),
-                               **config['wikis'][wiki])
+                               mw_host=wikimedia.host_from_dbname(wiki))
 
     return bp
