+++ wikilabels/wsgi/static/css/workspace.css
@@ -5,6 +5,7 @@
   padding: 1ex 1em;
   background: #fff;
   box-shadow: 0em 0em 1em rgba(0,0,0,.5);
+  display: none;
 }
 .wikilabels-workspace.fullscreen {
   position: absolute;
